♻️ : 모달 컴포넌트 타입스크립트 전환 및 Prop 네이밍 수정
